// Ladesar-Organics-backend/src/types.ts
export interface Address {
  id: string;
  fullName: string;
  phone: string;
  street: string;
  city: string;
  state: string;
  pincode: string;
  isDefault?: boolean;
  type: 'Home' | 'Work' | 'Other';
}

export type UserAddress = Address;

export type OrderStatus = 'Placed' | 'Processing' | 'Packed' | 'Shipped' | 'Out for Delivery' | 'Delivered' | 'Cancelled';

export interface OrderTimelineItem {
  status: OrderStatus;
  timestamp: string;
  description: string;
  completed: boolean;
}

export interface Order {
  id: string;
  orderNumber: string;
  createdAt: string;
  items: {
    productId: string;
    productName: string;
    productImage: string;
    variantSize: string;
    unitPrice: number;
    quantity: number;
    total: number;
  }[];
  shippingAddress: Address;
  paymentMethod: 'Razorpay' | 'UPI / PhonePe' | 'Credit / Debit Card' | 'Net Banking' | 'Cash on Delivery' | 'Wallet';
  paymentStatus: 'Paid' | 'Pending' | 'COD';
  transactionId?: string;
  subtotal: number;
  discountAmount: number;
  couponApplied?: string;
  shippingFee: number;
  taxAmount: number;
  totalAmount: number;
  status: OrderStatus;
  estimatedDelivery: string;
  timeline: OrderTimelineItem[];
  customerEmail: string;
  customerPhone: string;
}